requirements.txt
praw
pyyaml
nltk
airflow
pyspark